+++ larky/src/test/resources/META-INF/services/com.verygood.security.larky.modules.vgs.cerebro.text.analyzer.spi.TextPIIAnalyzer
@@ -1,1 +1,1 @@
-com.verygood.security.larky.modules.vgs.cerebro.text.analyzer.analyzers.DefaultTextPIIAnalyzer
+com.verygood.security.larky.modules.vgs.cerebro.text.analyzer.analyzers.InMemoryTextPIIAnalyzer
